Move echo() into echo.py
